@@ -1,34 +1,35 @@
 #pragma once
 #include <string>
-
-
+#include <string_view>
 
 namespace network {
+
+const std::string DELIMITERS = "[\\.,:-]";
 
 class ARP {
 
 public:
-  explicit ARP(const std::string& mask);
+    explicit ARP(std::string_view mask);
 
- const std::string& Address() const;
+    const std::string& Address() const;
 
 private:
-  std::string address_;
-  bool IsValid(const std::string& mask) const;
+    std::string address_;
 
-  std::string Parse(const std::string& mask) const;
+    bool IsValid(std::string_view mask) const;
 
-  template <typename Iterator>
-  std::string Join(Iterator first, Iterator last) const {
+    std::string Parse(std::string mask) const;
 
-    std::string result;
-    for (; first != last; ++first) {
-      result += *first;
+    template <typename Iterator>
+    std::string Join(Iterator first, Iterator last) const {
+
+        std::string result;
+        for (; first != last; ++first) {
+            result += *first;
+        }
+        return result;
     }
-    return result;
-  }
 
-  const std::string DELIMITERS = "[\\.,:-]";
 };
 
 } // namespace network
